requirements.txt
flask==3.1.3
python-dotenv==1.2.2
gunicorn==23.0.0
playwright==1.49.0
